+++ src/ImuSensor_TEST.cc
@@ -17,6 +17,7 @@
 #include <gtest/gtest.h>
 #include <sdf/sdf.hh>
 
+#include <ignition/math/Helpers.hh>
 #include <ignition/msgs.hh>
 #include <ignition/sensors/Export.hh>
 #include <ignition/sensors/Manager.hh>
@@ -27,13 +28,13 @@
 // These values come from the Rotors default values:
 // https://github.com/ethz-asl/rotors_simulator/blob/513bb92da0c1a0c968bdc679dffc8fe7d77de918/rotors_gazebo_plugins/include/rotors_gazebo_plugins/gazebo_imu_plugin.h#L40
 static constexpr double kDefaultAdisGyroscopeNoiseDensity =
-    2.0 * 35.0 / 3600.0 / 180.0 * M_PI;
+    2.0 * 35.0 / 3600.0 / 180.0 * IGN_PI;
 static constexpr double kDefaultAdisGyroscopeRandomWalk =
-    2.0 * 4.0 / 3600.0 / 180.0 * M_PI;
+    2.0 * 4.0 / 3600.0 / 180.0 * IGN_PI;
 static constexpr double kDefaultAdisGyroscopeBiasCorrelationTime =
     1.0e+3;
 static constexpr double kDefaultAdisGyroscopeTurnOnBiasSigma =
-    0.5 / 180.0 * M_PI;
+    0.5 / 180.0 * IGN_PI;
 static constexpr double kDefaultAdisAccelerometerNoiseDensity =
     2.0 * 2.0e-3;
 static constexpr double kDefaultAdisAccelerometerRandomWalk =